Move some parts to coco utils
