+++ deeplake/auto/unstructured/coco/constants.py
@@ -1,10 +1,11 @@
+from typing import Dict
 from deeplake.constants import UNSPECIFIED
 
 DEFAULT_GENERIC_TENSOR_PARAMS = {"htype": UNSPECIFIED, "sample_compression": None}
 DEFAULT_IMAGE_TENSOR_PARAMS = {"htype": "image", "sample_compression": "jpeg"}
 
 # Contains default kwargs for the tensors created from each of COCO keys
-DEFAULT_COCO_TENSOR_PARAMS = {
+DEFAULT_COCO_TENSOR_PARAMS: Dict[str, Dict] = {
     "segmentation": {
         "htype": "polygon",
         "sample_compression": None,
