+++ deeplake/tests/dummy_data/tests_auto/coco/annotations2.json
@@ -155,24 +155,6 @@
       {
         "segmentation": [
           [
-            192.81, 247.09, 199.36, 242.5, 205.92, 235.95, 214.44, 231.36, 226.24,
-            228.73, 235.42, 230.04, 241.97, 233.32, 245.91, 227.42, 252.46, 224.8,
-            262.95, 238.57, 267.54, 245.78, 264.92, 251.68, 265.58, 256.27,
-            257.71, 245.78, 250.5, 230.7, 248.53, 238.57, 243.94, 243.16, 234.11,
-            247.09, 226.9, 249.06, 227.55, 254.3, 224.27, 258.23, 219.03, 254.3,
-            219.03, 249.06
-          ]
-        ],
-        "area": 1035.749000000001,
-        "iscrowd": 0,
-        "image_id": 1268,
-        "bbox": [192.81, 224.8, 74.73, 33.43],
-        "category_id": 16,
-        "id": 42986
-      },
-      {
-        "segmentation": [
-          [
             264.31, 131.24, 263.81, 140.27, 181.67, 141.44, 124.77, 141.78,
             125.77, 135.25, 126.78, 130.9, 128.95, 126.22, 143.0, 125.22, 149.53,
             128.56, 148.86, 130.24, 213.1, 130.74, 216.44, 126.55, 237.52, 124.71,
